Extract mind map detail file components
